+++ doc/release_notes.rst
@@ -11,6 +11,8 @@ Release Notes
 .. Upcoming Release
 .. ================
 
+* Feature: Introduce a new method to overwrite costs (https://github.com/PyPSA/pypsa-eur/pull/1752). Modifications to the default technico-economic assumptions can now be enabled in the configuration using the `costs:custom_costs` setting (default value: `false`). When enabled, modifications specified in the file `data/custom_costs.csv` are applied to the default `resources/costs_{planning_horizons}.csv` files. The implementation maintains backwards compatibility with existing approaches via `costs:overwrites` and `costs:capital_cost`/`costs:marginal_cost` parameters.
+
 * Allow district heating potentials to be optionally specified as a dictionary as an alternative to a float, with country codes as keys of the parameter `district_heating: potential`.
 
 * Fix: Adjusted the `modify_attribute` function in `prepare_network` to align with changes in the PyPSA API.